make type check works
