+++ website/app/blog/page.tsx
@@ -1,48 +1,10 @@
-import { FC } from 'react';
-import { getAllBlogs } from '@all-blogs';
-import { asArray, BlogCardList, Heading, HeroSection, Newsletter, TagList } from '@components';
+import { BlogPage } from './blog-page';
 
 export const metadata = {
   title: 'Blog',
   description: 'Announcements about our Open-Source projects',
 };
 
-function extractRelevantTags(articles) {
-  const allTags = articles.flatMap(article => article.tags || []);
-  const map = Object.create(null);
-  for (const tag of allTags) {
-    map[tag] = map[tag] == null ? 0 : map[tag];
-    map[tag] += 1;
-  }
-  const sorted = Object.entries(map)
-    .sort((a, b) => b[1] - a[1])
-    .slice(0, 10);
-  if (sorted.every(([tagName]) => tagName !== 'codegen')) {
-    sorted.unshift(['codegen', 0]);
-  }
-  if (sorted.every(([tagName]) => tagName !== 'envelop')) {
-    sorted.unshift(['envelop', 0]);
-  }
-  return sorted.slice(0, 10) as unknown as string[];
+export default function Page() {
+  return <BlogPage />;
 }
-
-const BlogPage: FC<{ tag: string }> = async ({ tag }) => {
-  const allBlogs = await getAllBlogs();
-  const allTags = extractRelevantTags(allBlogs);
-  const articles = tag ? allBlogs.filter(article => asArray(article.tags).includes(tag)) : allBlogs;
-
-  return (
-    <>
-      <HeroSection>
-        <Heading>The Guild's blog</Heading>
-      </HeroSection>
-      <div className="nextra-container">
-        <TagList tags={allTags} withCount className="mb-20 mt-10" />
-        {!tag && <Newsletter />}
-        <BlogCardList articles={articles} />
-      </div>
-    </>
-  );
-};
-
-export default BlogPage;
