+++ website/package.json
@@ -5,7 +5,6 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "//prebuild": "pnpm types:check",
     "analyze": "ANALYZE=true pnpm build",
     "build": "next build && SITE_URL=https://the-guild.dev next-sitemap",
     "dev": "next --turbopack",
